+++ app/src/main/java/com/example/driverchecker/database/PartialEntity.kt
@@ -22,10 +22,10 @@ data class PartialEntity (
     @ColumnInfo(name = "confidence") val confidence: Float,
     @ColumnInfo(name = "index") val index: Int,
     @ColumnInfo(name = "group") val group: String,
-    @ColumnInfo(name = "evaluation_id") val evaluationId: Int,
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @ColumnInfo(name = "evaluation_id") val evaluationId: Long,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
 ) {
-    constructor(partialResult: IImageDetectionOutput<String>, outputIndex: Int, evalId: Int) : this (
+    constructor(partialResult: IImageDetectionOutput<String>, outputIndex: Int, evalId: Long) : this (
         confidence = partialResult.confidence,
         index = outputIndex,
         group = partialResult.groups.keys.first(),
